Added partial castling functionality.
Added functionality for white to castle on the right.
Added functionality for black to castle on the right.
Added validation to not allow castling if white Rook or white King has been moved.
Added validation to not allow castling if black Rook or black King has been moved.
Optimized gameplay.